requirements.txt
fpdf
python-dotenv
google-genai